+++ src/error.rs
@@ -91,6 +91,7 @@ pub enum Error {
     InvalidIdentifier(String),
     SuggestRawIdentifier(String),
     ExceededRecursionLimit,
+    ExpectedRawValue,
 }
 
 impl fmt::Display for SpannedError {
@@ -250,6 +251,7 @@ impl fmt::Display for Error {
                 identifier, identifier
             ),
             Error::ExceededRecursionLimit => f.write_str("Exceeded recursion limit, try increasing the limit and using `serde_stacker` to protect against a stack overflow"),
+            Error::ExpectedRawValue => f.write_str("Expected a `ron::value::RawValue`"),
         }
     }
 }