+++ src/parse.rs
@@ -359,7 +359,7 @@ impl<'a> Bytes<'a> {
         }
     }
 
-    pub fn bytes(&self) -> &[u8] {
+    pub fn bytes(&self) -> &'a [u8] {
         self.bytes
     }