+++ src/ser/mod.rs
@@ -14,6 +14,7 @@ use crate::{
     },
 };
 
+mod raw;
 #[cfg(test)]
 mod tests;
 mod value;
@@ -572,6 +573,10 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
     where
         T: ?Sized + Serialize,
     {
+        if name == crate::value::raw::RAW_VALUE_TOKEN {
+            return value.serialize(raw::RawValueSerializer::new(self));
+        }
+
         if self.extensions().contains(Extensions::UNWRAP_NEWTYPES) || self.newtype_variant {
             self.newtype_variant = false;